requirements.txt
flask
flask-cors
selenium
openai
python-dotenv
requests